requirements.txt
numpy>=2.1
opencv-python>=4.10
matplotlib>=3.9
imageio>=2.35
imageio-ffmpeg>=0.5
Pillow>=10.4
python-dotenv>=1.0
fastapi>=0.115
google-cloud-storage>=2.18
firebase-admin>=6.5
uvicorn>=0.34.0
